requirements.txt
pyrogram==2.0.106
pytgcalls==3.0.1
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0
tgcrypto==1.2.5
